+++ src/operator/tensor/indexing_op.cu
@@ -435,4 +435,3 @@ NNVM_REGISTER_OP(_scatter_set_nd)
 .set_attr<FCompute>("FCompute<gpu>", ScatterSetNDForward<gpu>);
 }  // namespace op
 }  // namespace mxnet
-
